requirements.txt
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
idna==3.7
numpy==1.24.4
packaging==24.1
pandas==2.0.3
proto-plus==1.24.0
protobuf==5.27.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
